+++ SpawnProcess.h
@@ -16,9 +16,9 @@ public:
     int Spawn(const wstring& strCmd, const wstring& strWorkingDir = wstring()) noexcept;
     int KillProcess() noexcept;
     bool StillSpawning() noexcept;
-    int ReadFromSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
-    int ReadErrFromSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
-    int WriteToSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
+    size_t ReadFromSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
+    size_t ReadErrFromSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
+    size_t WriteToSpawn(unsigned char* const pBuffer, const uint32_t nBufSize) noexcept;
     void CloseWritePipe() noexcept;
     void AddEnvironment(const wstring& strEnvironment) noexcept { m_vstrEnvironment.push_back(wstring_convert<codecvt_utf8<wchar_t>, wchar_t>().to_bytes(strEnvironment)); }
     void AddEnvironment(const string& strEnvironment) noexcept { m_vstrEnvironment.push_back(strEnvironment); }
